capture image stack and project on three dimensions
This program will only run the camera to take image stack. Hence before
running this program, the tunable lens and scanning mirror should be
turn on and reach thermo stability.

This program will continue capturing image stacks and project on 3 axis
quickly (1~2Hz) for alignment. Or
capture nrCycles of images and do interpolation to correct the axial
position of the image. Then produce comparison of raw/processed image
projections.
offsetFrames need to be provided for accurate position.